by-dataframe/polars/requirements.txt
adbc-driver-postgresql<1.9
connectorx>=0.4.5a1
pandas<2.4
polars<1.36
pyarrow<23
sqlalchemy-cratedb>=0.42.0.dev2
